hterm: a11y options page button in scrollport
The recent a11y review of terminal uncovered that the options
page is currently not discoverable for users of chromevox.
For terminal system app, it doesn't seem that it is always
possible to use chromevox to navigate to the frame and use
the menu.  This adds an a11y options button in scrollport
below the page down button.

Added Terminal onOpenOptionsPage where clients can register
a function to handle opening the options page.

Chromevox must have changed again how it selects elements,
so reverted crrev.com/c/161286.

Change-Id: I65501b18ec1fc2fff54de976e4337e38871f3c21
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2133526
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>