hterm: add padding with new screen-padding-size property
Having screen padding leaves space around the border of the terminal
and makes it easier to use the mouse to select text at the start of
a row.

Default screen padding size set to 8 pixels.

The padding is created by setting rowNodes inside screen at
a fixed position of the padding size.  It is not possible to use
CSS padding or margin since the screen and rowNodes elements
are set as fixed position with 100% width / height.

By putting the padding between screen and rowNodes, the
scrollbar remains on the edge of the window which is
preferable.

Bug: 267654
Change-Id: Ibd0915505ab433e33546a23dc7dad884e5feed12
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2143311
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>