eslint: disable l/I variable usage by default
These are hard to distinguish depending on the font, so block them by
default.  We have a few limited uses which (hopefully) should be clear
based on surrounding context that we'll still allow, but people have
to explicitly opt-in to it.

Some code in terminal is cleaned up to make it more readable as it
was using too many single letter variables including "l".

Change-Id: I8ab4e70ea5db9c8c90f7d6d14cf316af2cbe4226
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2144946
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>