libdot: mdlint: new tool for linting markdown files
This just sets up the framework for linting the files, but doesn't
turn on any rules.  We'll need to clean files up as we go first.

We have to tweak the style with a few *** blocks to workaround a
bug in the tool -- it hangs when using italics in them.

Change-Id: Ia43b2117b2b202740c51b24aab76ae1a6847a86c
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2245041
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>