hterm: fix VT region scrolling
Fix the VT csi-r DECSTBM parsing to correctly handle when top parameter
is not included or is set to zero.  We were previously doing a 1-based
to 0-based conversion and calling setVTScrollRegion with -1, when it
should be null.

Fixed scrolling behavior to scroll rows from screen to scrollback when
|vtScrollTop_| is not set. When only the VT bottom region is set, we
should continue to scroll rows into scrollback. Added new insertRow_ to
support the case where |vtScrollBottom_| is set, but |vtScrollTop_| is
not and new rows need to be added in the middle of the screen.

Bug: 1106718
Change-Id: Ia5ba96bb57ff7c2a33fb29de22eb9247f86d1eda
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2315460
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>