hterm: change Terminal constructor API to an options object
To prepare for adding more options to the Terminal initialization,
change it from taking a single optional string to an optional object
of multiple options.  This makes it easier to pass in subsets w/out
having to pass null/undefined for earlier fields.

This breaks the API a bit, but no way around it for long term health.

Change-Id: Iaa78552f163ec173881618b74fe2fd37fccb2385
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2494361
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>