hterm: disable zoom gestures
When scrolling with the mouse wheel and holding the Control key,
Chrome & Firefox will start zooming the page.  If we have content
to scroll, we we're defeating that behavior already, but if we
didn't, we let the browser take over and it'd trigger zooming.
Let's always disable when Control is held to be consistent.

This also cleans up touchpad pinch gestures which generate fake
scroll wheel events as if the Control key were held.

Bug: chromium:896490, chromium:1149023
Change-Id: I652969cc6991b8d5ac7eb19b4efbc69448d28fa3
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2536778
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>