hterm: notifications: keep it dynamically centered
Currently any overlays shown are centered based on the size of the
terminal when initialized.  If the overlay stays up long enough for
the terminal to resize, it does not shift to the new center.  This
isn't a big deal for most overlays as they're short lived (seconds),
but for anything longer, the UX is bad.

Switch to CSS calc() so that we keep the dialog centered.  We'll
assume for now the overlay itself doesn't change size, so we'll use
its initial dimensions to keep it centered.  We can revisit that
later (perhaps with CSS vars) if the need arises, but even in the
short term, it wouldn't be a big deal for it to be slightly off.

Along those lines, we don't bother including the width of the
scrollbar anymore to help simplify.  That normally only takes up
<16 px, and no one is going to notice a <8 px error.

Bug: https://crbug.com/609303
Change-Id: Iec986200381d204effe5db383bcaf3f8a567eba4
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2563928
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>