hterm: notifications: pull out into dedicated file
We want to do more with this interface, but it'll be a lot cleaner
if we have its own file to hold everything.  Rip it out of Terminal
and into a new dedicated class.

There should be no API changes in here -- the Terminal show & hide
overlay methods should behave the same.  We refactor the code a bit
to make it easier to unittest (and we add those tests).

Bug: https://crbug.com/609303
Change-Id: I0e50221994a763e40694d103187992bb28e998e5
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2563443
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>